pyproject.toml
[project]
name = "tierlist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pick>=0.6.7",
    "pillow>=12.1.1",
    "pylast>=7.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
]
